refactor(llm): layer the kernel families as library modules
Split the two monolithic kernel files into module directories with one
layer per file and module-level documentation, so each kernel family
reads the same way:

- scan/ (was 1,860 lines flat): mod.rs (public entry point, backend
  extension trait, checkpointing contract), reference.rs (differentiable
  tensor-op oracle + CPU impl), autodiff.rs (first-order node), gpu.rs
  (CubeCL kernels + dispatch), tests.rs (parity at adversarial
  geometries).
- linear_cross_entropy/ (was 1,074 lines flat): same five layers.
- model/mod.rs documents the crate's layering (model / kernel families /
  runtime glue) so the structure is discoverable from the front door.

Pure code motion plus imports and module docs - no logic, signature, or
kernel changes (Metal suite 24/24, hermes-train suite, and clippy on both
feature sets green; the compiler proves the moves).
